+++ src/fsrs/fsrs.py
@@ -16,10 +16,10 @@ from datetime import datetime, timezone, timedelta
 from copy import deepcopy
 from math import inf
 from typing import Any
-from enum import IntEnum
 import random
 
-from fsrs.fsrs_v5 import Rating, FSRSv5
+from fsrs.fsrs_v5 import FSRSv5
+from fsrs.models import Card, Rating, State, ReviewLog
 
 
 FUZZ_RANGES = [
@@ -27,201 +27,6 @@ FUZZ_RANGES = [
     {"start": 7.0, "end": 20.0, "factor": 0.1},
     {"start": 20.0, "end": inf, "factor": 0.05},
 ]
-
-
-class State(IntEnum):
-    """
-    Enum representing the learning state of a Card object.
-    """
-
-    Learning = 1
-    Review = 2
-    Relearning = 3
-
-
-class Card:
-    """
-    Represents a flashcard in the FSRS system.
-
-    Attributes:
-        card_id (int): The id of the card. Defaults to the epoch miliseconds of when the card was created.
-        state (State): The card's current learning state.
-        step (int | None): The card's current learning or relearning step or None if the card is in the Review state.
-        stability (float | None): Core mathematical parameter used for future scheduling.
-        difficulty (float | None): Core mathematical parameter used for future scheduling.
-        due (datetime): The date and time when the card is due next.
-        last_review (datetime | None): The date and time of the card's last review.
-    """
-
-    card_id: int
-    state: State
-    step: int | None
-    stability: float | None
-    difficulty: float | None
-    due: datetime
-    last_review: datetime | None
-
-    def __init__(
-        self,
-        card_id: int | None = None,
-        state: State = State.Learning,
-        step: int | None = None,
-        stability: float | None = None,
-        difficulty: float | None = None,
-        due: datetime | None = None,
-        last_review: datetime | None = None,
-    ) -> None:
-        if card_id is None:
-            # epoch miliseconds of when the card was created
-            card_id = int(datetime.now(timezone.utc).timestamp() * 1000)
-        self.card_id = card_id
-
-        self.state = state
-
-        if self.state == State.Learning and step is None:
-            step = 0
-        self.step = step
-
-        self.stability = stability
-        self.difficulty = difficulty
-
-        if due is None:
-            due = datetime.now(timezone.utc)
-        self.due = due
-
-        self.last_review = last_review
-
-    def to_dict(self) -> dict[str, int | float | str | None]:
-        """
-        Returns a JSON-serializable dictionary representation of the Card object.
-
-        This method is specifically useful for storing Card objects in a database.
-
-        Returns:
-            dict: A dictionary representation of the Card object.
-        """
-
-        return_dict = {
-            "card_id": self.card_id,
-            "state": self.state.value,
-            "step": self.step,
-            "stability": self.stability,
-            "difficulty": self.difficulty,
-            "due": self.due.isoformat(),
-            "last_review": self.last_review.isoformat() if self.last_review else None,
-        }
-
-        return return_dict
-
-    @staticmethod
-    def from_dict(source_dict: dict[str, Any]) -> "Card":
-        """
-        Creates a Card object from an existing dictionary.
-
-        Args:
-            source_dict (dict[str, Any]): A dictionary representing an existing Card object.
-
-        Returns:
-            Card: A Card object created from the provided dictionary.
-        """
-
-        card_id = int(source_dict["card_id"])
-        state = State(int(source_dict["state"]))
-        step = source_dict["step"]
-        stability = (
-            float(source_dict["stability"]) if source_dict["stability"] else None
-        )
-        difficulty = (
-            float(source_dict["difficulty"]) if source_dict["difficulty"] else None
-        )
-        due = datetime.fromisoformat(source_dict["due"])
-        last_review = (
-            datetime.fromisoformat(source_dict["last_review"])
-            if source_dict["last_review"]
-            else None
-        )
-
-        return Card(
-            card_id=card_id,
-            state=state,
-            step=step,
-            stability=stability,
-            difficulty=difficulty,
-            due=due,
-            last_review=last_review,
-        )
-
-
-class ReviewLog:
-    """
-    Represents the log entry of a Card object that has been reviewed.
-
-    Attributes:
-        card (Card): Copy of the card object that was reviewed.
-        rating (Rating): The rating given to the card during the review.
-        review_datetime (datetime): The date and time of the review.
-        review_duration (int | None): The number of miliseconds it took to review the card or None if unspecified.
-    """
-
-    card: Card
-    rating: Rating
-    review_datetime: datetime
-    review_duration: int | None
-
-    def __init__(
-        self,
-        card: Card,
-        rating: Rating,
-        review_datetime: datetime,
-        review_duration: int | None = None,
-    ) -> None:
-        self.card = deepcopy(card)
-        self.rating = rating
-        self.review_datetime = review_datetime
-        self.review_duration = review_duration
-
-    def to_dict(self) -> dict[str, dict[str, Any] | int | str | None]:
-        """
-        Returns a JSON-serializable dictionary representation of the ReviewLog object.
-
-        This method is specifically useful for storing ReviewLog objects in a database.
-
-        Returns:
-            dict: A dictionary representation of the Card object.
-        """
-
-        return_dict = {
-            "card": self.card.to_dict(),
-            "rating": self.rating.value,
-            "review_datetime": self.review_datetime.isoformat(),
-            "review_duration": self.review_duration,
-        }
-
-        return return_dict
-
-    @staticmethod
-    def from_dict(source_dict: dict[str, Any]) -> "ReviewLog":
-        """
-        Creates a ReviewLog object from an existing dictionary.
-
-        Args:
-            source_dict (dict[str, Any]): A dictionary representing an existing ReviewLog object.
-
-        Returns:
-            ReviewLog: A ReviewLog object created from the provided dictionary.
-        """
-
-        card = Card.from_dict(source_dict["card"])
-        rating = Rating(int(source_dict["rating"]))
-        review_datetime = datetime.fromisoformat(source_dict["review_datetime"])
-        review_duration = source_dict["review_duration"]
-
-        return ReviewLog(
-            card=card,
-            rating=rating,
-            review_datetime=review_datetime,
-            review_duration=review_duration,
-        )
 
 
 class Scheduler:
